requirements.txt
numpy
python-dotenv
tk